requirements.txt
Brotli==1.1.0
Deprecated==1.2.14
Django==5.1
Faker==27.0.0
Jinja2==3.1.4
Markdown==3.6
MarkupPy==1.14
MarkupSafe==2.1.5
PyYAML==6.0.2
Pygments==2.18.0
Willow==1.8.0
anyascii==0.3.2
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==4.1.0
boto3==1.34.159
botocore==1.34.159
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
confusable-homoglyphs==3.3.1
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
cryptography==43.0.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
dj-stripe==2.8.4
django-after-response==0.2.2
django-allauth==64.0.0
django-appconf==1.0.6
django-ckeditor==6.7.1
django-cogwheels==0.3
django-colorful==1.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.3
django-cryptography-django5==2.2
django-debug-toolbar==4.4.6
django-dynamic-image==1.4.0
django-el-pagination==4.0.0
django-extensions==3.2.3
django-filter==24.3
django-hijack==3.6.0
django-honeypot==1.2.1
django-imagekit==5.0.0
django-import-export==4.1.1
django-ipware==7.0.1
django-js-asset==2.2.0
django-modelcluster==6.3
django-multiselectfield==0.1.13
django-permissionedforms==0.1
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-recurrence==1.11.1
django-registration==3.4
django-richtextfield==1.6.2
django-sendgrid-v5==1.2.3
django-social-share==2.3.0
django-sql-explorer==5.1.1
django-storages==1.14.4
django-tables2==2.7.0
django-taggit==5.0.1
django-timezone-field==7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docutils==0.21.2
draftjs-exporter==5.0.0
enmerkar==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
filetype==1.2.0
fonttools==4.53.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
html2docx==1.6.0
html5lib==1.1
icalendar==5.0.13
idna==3.7
importlib_metadata==8.0.0
jmespath==1.0.1
l18n==2021.3
laces==0.1.1
logfire==0.50.1
lxml==5.3.0
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.56
markdown-it-py==3.0.0
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-jinja2==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-sqlite3==0.47b0
opentelemetry-instrumentation-system-metrics==0.47b0
opentelemetry-instrumentation-urllib3==0.47b0
opentelemetry-instrumentation-urllib==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
oscrypto==1.3.0
packaging==24.1
phonenumbers==8.13.43
pilkit==3.0
pillow==10.4.0
pillow_heif==0.18.0
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pyHanko==0.25.1
pycparser==2.22
pydyf==0.11.0
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-http-client==3.3.7
python-ipware==3.0.0
python-slugify==8.0.4
python-webpack-boilerplate==1.0.3
pytz==2024.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
sendgrid==6.11.0
setuptools==72.1.0
six==1.16.0
sorl-thumbnail==12.10.0
soupsieve==2.6
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
stripe==4.2.0
svglib==1.5.1
tablib==3.5.0
telepath==0.3.1
text-unidecode==1.3
texttable==1.7.0
tinycss2==1.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
wagtail-color-panel==1.5.0
wagtail-markdown==0.12.1
wagtail-modeladmin==2.0.0
wagtail==6.2
wagtailmenus==4.0.1
weasyprint==62.3
webencodings==0.5.1
wheel==0.44.0
whitenoise==6.7.0
wrapt==1.16.0
xhtml2pdf==0.2.16
xlrd==2.0.1
xlwt==1.3.0
zipp==3.20.0
zopfli==0.2.3
vendor/puput-2.0.0.zip
vendor/wagtail-django-recaptcha-1.0.zip
vendor/wagtail-seo-2.4.1.zip
